reformatted code, broken down to smaller parts, more control to hover/add points
